MergeGuitarTuner - searchNearestMatch added

+++ app/src/main/java/com/awaj/BinarySearch.java
@@ -10,11 +10,11 @@ public class BinarySearch {
      * @param: float[] , float
      * @return: int
      */
-    public static int searchNearestValue(float[] data, float value){
+    public static int searchNearestValue(double[] data, float value){
         float distance =10000;
         int index=0;
         for(int i=0;i<data.length;i++){
-            float cDistance = Math.abs(data[i]-value);
+            float cDistance =(float) Math.abs(data[i]-value);
             if(cDistance<=distance){
                 index =i;
                 distance = cDistance;
